scripts/fetcher_tests/requirements.txt
avro-python3==1.10.2
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.5
urllib3==2.5.0